Create File_organise_Segment.py
When files are deconvolved with imagej, the deconvolved ones are within the directory specific to the animal id, lets say 18. Now similar folder structure is created by created Deconvolved_ims directory inside which there is a dir for each animal id and inside which the images are divided into col1a1 subdir and hunu subdir.