Update Stats_Calculation.py
there was a bug that caused a formation of doubles or even triples of values due to the way hunu images were saved. thus, added   files_subdir_h = glob.glob(subdir_p + '*WS.png') so that only the watershedded ones are included automatically. the if statement after that ( if 'WS' in file:) on the line 224 is redundant and will be removed later